Update dependencies
- completalbe-futures 0.3.3 -> 0.3.4
- dropwizard-core 2.0.13 -> 2.0.15
- Fastutil 8.4.2 -> 8.4.3
- gax-grpc 1.60.0 -> 1.60.1
- gRPC 1.33.0 -> 1.33.1
- grpc-kotlin-stub 0.2.0 -> 0.2.1
- io.dropwizard.metrics 4.1.13 -> 4.1.15
- Jackson 2.11.2 -> 2.11.3
- Micrometer 1.5.5 -> 1.6.1
- org.bouncycastle 1.66 -> 1.67
- Reactor 3.3.10.RELEASE -> 3.4.0
- reactor-kotlin-extensions 1.0.2.RELEASE -> 1.1.0
- Spring Boot 2.3.4.RELEASE -> 2.4.0
- spring-web 5.2.9.RELEASE -> 5.3.1.RELEASE
- tomcat-embed-core 9.0.39 -> 9.0.40

- Examples and testings
  - akka-actor-typed_2.13 2.6.9 -> 2.6.10
  - akka-grpc-runtime_2.13 1.0.1 -> 1.0.2
  - Checkstyle 8.36.2 -> 8.37
  - Dagger 2.29.1 -> 2.30.1
  - Eureka 1.10.7 -> 1.10.10
  - Finagle 20.9.0 -> 20.10.0
  - Gradle 6.7 -> 6.7.1
  - httpclient 4.5.12 -> 4.5.13
  - jmh-gradle-plugin 0.5.0 -> 0.5.2
  - json-unit 2.19.0-> 2.21.0
  - kotlin-allopen 1.4.10 -> 1.4.20
  - mockito-core 3.5.15 -> 3.6.0
  - monix-reactive_2.13 3.2.2+45-5c6c8b9e -> 3.3.0
  - proguard-gradle 7.0.0 -> 7.0.1

- Removed `application-testbed.yml` in example because Spring now sees the `application.yml` in the Jar first.
- boot-autoconfigure and boot-actuator-autoconfigure does not have starter dependency anymore.
  - boot-webflux-autoconfigure has stater dependency though because there are no autoconfigure dependency for webflux.
