+++ build.gradle
@@ -6,7 +6,7 @@ plugins {
     id 'com.google.osdetector' version '1.6.2' apply false
     // If you want to change `org.jetbrains.kotlin.jvm` version,
     // you also need to change `org.jetbrains.kotlin:kotlin-allopen` version in `dependencies.yml`.
-    id 'org.jetbrains.kotlin.jvm' version '1.4.10' apply false
+    id 'org.jetbrains.kotlin.jvm' version '1.4.20' apply false
 
     id 'net.ltgt.errorprone' version '1.2.1' apply false
 }
