+++ examples/spring-boot-minimal-kotlin/src/test/kotlin/example/springframework/boot/minimal/kotlin/HelloApplicationIntegrationTest.kt
@@ -8,9 +8,7 @@ import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.boot.test.context.SpringBootTest
-import org.springframework.test.context.ActiveProfiles
 
-@ActiveProfiles("testbed")
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.NONE)
 class HelloApplicationIntegrationTest(@Autowired server: Server) {
