--- examples/spring-boot-minimal-kotlin/src/test/resources/application-testbed.yml
@@ -1,10 +1,0 @@
-# Disable the Web Server.
-# See https://docs.spring.io/spring-boot/docs/current/reference/html/howto-embedded-web-servers.html.
-spring.main.web-application-type: none
-
-# Additional server ports specified with Armeria configuration.
-armeria:
-  ports:
-  - port: 0
-    protocols:
-    - http
