+++ examples/spring-boot-minimal/src/main/resources/config/application.yml
@@ -5,6 +5,6 @@ spring.main.web-application-type: none
 # Additional server ports specified with Armeria configuration.
 armeria:
   ports:
-  - port: 8080
+  - port: 0 # Specify the port you want
     protocols:
     - http
