--- examples/spring-boot-tomcat/src/test/resources/application-testbed.yml
@@ -1,7 +1,0 @@
-# Prevent the embedded Tomcat from opening a TCP/IP port.
-server.port: -1
----
-armeria:
-  ports:
-    - port: 0
-      protocol: HTTP
