+++ spring/boot2-actuator-autoconfigure/src/test/resources/application-secureTest.yml
@@ -1,3 +1,8 @@
+# Use-legacy-processing to use this application-testbed.yml file instead of application.yml
+spring:
+  config:
+    use-legacy-processing: true
+
 armeria:
   ports:
     - port: 0
