+++ spring/boot2-starter/build.gradle
@@ -1,3 +1,4 @@
 dependencies {
     api project(':spring:boot2-autoconfigure')
+    api 'org.springframework.boot:spring-boot-starter'
 }
