+++ spring/boot2-webflux-autoconfigure/src/main/java/com/linecorp/armeria/spring/web/reactive/ArmeriaClientHttpRequest.java
@@ -106,6 +106,16 @@ final class ArmeriaClientHttpRequest extends AbstractClientHttpRequest {
         return uri;
     }
 
+    @Nullable
+    @SuppressWarnings("override")
+    public <T> T getNativeRequest() {
+        if (request == null) {
+            return null;
+        }
+        //noinspection unchecked
+        return (T) request;
+    }
+
     @Override
     public DataBufferFactory bufferFactory() {
         return factoryWrapper.delegate();
